Lots of changes, removed the node world structure, and replaced it with a normal x,y world with objects in it. this makes it easier tho notice a agent actually moving instead of teleporting from node to node
implemented the GOAP planner with a lot of changes to the GOAPAgent to support it.

Implemented the Elite::FSM to start on the state machine connected to it.

cleaned some code up and classes, moved a lot of members vars around, added comments.

fixed a lot of compile issues. started on FSMstates and transitions.